USB info fix + more Sysex request

+++ src/Components/UIelement.h
@@ -16,7 +16,7 @@ class UIelement
 public:
   UIelement();
   void renderAscii(char ascii, u8 xy, u32 color);
-  void renderHalfHeightNum(u8 num, u8 xy, u32 color, u32 sec_color);
+  void renderHalfHeightNum(u16 num, u8 xy, u32 color, u32 sec_color);
   void renderHalfHeightDigit(u8 num, u8 xy, u32 color);
   u8 binary8bitInput(u8 currentNum, u8 y, u32 color);
   u8 simple8bitInput(s16 currentNum, u8 y, u32 color);
